[ios, macos] Use integer to represent tile size in raster source

+++ platform/darwin/src/MGLRasterSource.h
@@ -8,7 +8,7 @@ NS_ASSUME_NONNULL_BEGIN
 @interface MGLRasterSource : MGLSource
 
 @property (nonatomic, readonly, copy) NSURL *URL;
-@property (nonatomic, readonly, assign) CGFloat tileSize;
+@property (nonatomic, readonly, assign) NSUInteger tileSize;
 @property (nonatomic, readonly, nullable) MGLTileSet *tileSet;
 
 - (instancetype)initWithSourceIdentifier:(NSString *)sourceIdentifier URL:(NSURL *)url tileSize:(CGFloat)tileSize;
